util: add a restore path option
Zenfs does not restore the files into the db path, For ex: If
"rocksdbtest/dbbench" is the db path from where backup was done,
zenfs uses the path "" is used to restore the files and the files
in db path which the application is referring to is same as previous.

Add option to zenfs to specify the path to restore the files.

$zenfs restore --zbd=nvme0n1 --path=/bkp/rocksdbtest/dbbench/ --restore_path=rocksdbtest/dbbench

Signed-off-by: Aravind Ramesh <aravind.ramesh@wdc.com>