util: fix zenfs list sub command to handle improper path argument
zenfs subcommand "list" fails to list ZenFS files when extra
forward slashes are present in the user provided path.
This works in the auxilary path, which is on a posix
filesystem which can handle extra slashes in the path.

Add format_path() helper function to parse user provided
path for listing files. This helper removes any duplicate
forward slashes in the path string and at the start of the
path string. Also check for returned error and report it.

Signed-off-by: Aravind Ramesh <aravind.ramesh@wdc.com>