Allow zenfs backup to be run on an already mounted ZBDs
There is no need for 'zenfs' utility to require exclusive access to the drive
when performing backups. Therefore, in the 'backup' command handler 'zbd' is now opened
and 'zenfs' is now mounted in read-only mode.
Please note, that files still being written may be backed up only partially, so it
remains user's responsibility to ensure that the operation they are performing is
safe.