util: allow backups of already mounted filesystems for testing purposes
Allow for backing up mounted file systems with a live writer
for testing purposes.

We need some form of snapshots to do this without
risking data corruption/loss, but this is good enough
for now to get zenfs compatible with the mysql/myrocks
mtr tests.

Do this by allowing for non-exclusive block device
access when running the backup command with the --force flag.
A warning is emitted and the the potential data corruption
is documented in the help.

Co-authored-by: Yura Sorokin <yura.sorokin@percona.com>
Signed-off-by: Hans Holmberg <hans.holmberg@wdc.com>