fs: store active extent start
If we store the current extent start and there is a crash
we know that this file wrote the data starting from
active extent start up to the zone write pointer.

We don't need to store the active zone as we can look it up
from extent_start_

TODO: this is backwards compatible, but will result in mount errors
for previous zenfs versions, so we need to step format version.

Signed-off-by: Hans Holmberg <hans.holmberg@wdc.com>