tests: print and store time spent for tests and test sets
It's useful to know how long a test runs, both for scheduling
testing and to compare different runs to look for regressions
and improvements in execution time.

So, print the time it takes to run each test and the total
time to run a whole test set in run.sh. Also store the time in the results
directory for comparing different runs.

Signed-off-by: Hans Holmberg <hans.holmberg@wdc.com>