fs, util: Introducing semi automatic versioning for ZenFS.
When building ZenFS through RocksDB the ZenFS version is updated by
its Makefile and contains the output of `git describe --abbrev=7 --dirty`.
This means that the version depends on the latest git tag and if
preceding commits exits the latest commit's short hash
(see (documentation)[https://git-scm.com/docs/git-describe]).

The version is regenerated if it differs from the last one.

The version string is persisted within the superblock.
A new util command `fs-info` was added that outputs the contents of the current
superblock which includes the ZenFS version information.

Suggested-by: Andreas Hindborg <andreas.hindborg@wdc.com>
Signed-off-by: Dennis Maisenbacher <dennis.maisenbacher@wdc.com>