fs: Add ZenFSMetrics to ZenFS
ZenFSMetrics is used to monitor some of the variables on ZenFS in real time.
In addition to this, two other classes have been added to metrics.h.
NoZenFSMetrics is an empty implementation of ZenFSMetrics and is used by default.
ZenFSMetricsLatencyGuard is used to count the execution time of a function,
and it reports the time between construction and destruction to Metrics.

Signed-off-by: Liu Ruicheng <sagitrs67@gmail.com>

Co-authored-by: Andreas Hindborg <andreas@metaspace.dk>

Signed-off-by: Hans Holmberg <hans.holmberg@wdc.com>
(Rebased the patch and fixed up a few non-debug compilation issues)