fs: avoid finishes in allocator by relaxing the life time diff requirement
If the write buffer is not big enough to fill a zone, the first-level ssts
will end up as many small files and cause the allocator to finish a lot of zones as the
write life time diff requirement cannot be met.

Avoid the finishes by allowing same-level-ssts to co-locate the same zones
when we're at the max active limit.

Signed-off-by: Hans Holmberg <hans.holmberg@wdc.com>