fs: Refactor and make the IO Zone allocator re-entrant
Make the IO Zone re-entrant in order parallelize zone allocation,
allowing the WAL files to get a zone immediately even though
an SST file might be waiting for a finish to be able to allocate
a new

In order to do this in a safe way, introduce a token-api for
active and open zones. Also give the allocator code a well-deserved
refactor makeover.

Signed-off-by: Hans Holmberg <hans.holmberg@wdc.com>