fs: mark up ZoneFiles as deleted in DeleteFile
As the zone files are shared pointers, readers
and migration threads might still hold a reference
to the file even though the file is deleted. That
is ok as the data is kept on disk until the last
reference has been removed.

We should not, however, migrate the data if the file has been
deleted so add a flag that indicates that the file
has been deleted.

Signed-off-by: Hans Holmberg <hans.holmberg@wdc.com>