fs: check if a zone file IsDeleted before syncing metadata
Files may be deleted or renamed at any time we migrate files in
a garabage-collection thread. By checking IsDeleted rather
than checking if the file name exists, we make sure that
we we won't write spurious update records for a file
that is deleted and/or renamed while being migrated.

This also fixes an issue where metadata is not written
when files are created (as the file was not yet part of
the file list, the metadata write was skipped).

Signed-off-by: Hans Holmberg <hans.holmberg@wdc.com>