fs: check IsDeleted during file migration
Files may be deleted or renamed at any time we migrate files in
a garabage-collection thread. By checking IsDeleted rather
than checking if the file name exists, we make sure that
we we won't write spurious update records for a file
that is deleted and/or renamed while being migrated.

Signed-off-by: Hans Holmberg <hans.holmberg@wdc.com>