fs: add DeleteFileNoLock()
DeleteFile() api deletes the said file while holding the files_mtx_.
Filesystem operations which are holding the files_mtx_ had to
relinquish it, if it needed to call DeleteFile(), then reacquire
it after DeleteFile was successful. This is not thread safe.
Introduce DeleteFileNoLock() which can be called by functions
holding the file lock.

Signed-off-by: Aravind Ramesh <aravind.ramesh@wdc.com>
Signed-off-by: Dennis Maisenbacher <dennis.maisenbacher@wdc.com>