test: Workaround bug in dbstress in rocksdb 6.29.3
dbstress in rocksdb 6.29.3 has a bug where it tries to create a directory with
name of `dbname`. Default `dbname` in dbstress is `rocksdbtest/dbstress`. As
directory `rocksdbtest` does not exist when dbstress tries to create
subdirectory `dbstress` an error is thrown. This patch fixes the issue by
setting `dbname` to a string without forwards slashes.

Signed-off-by: Andreas Hindborg <andreas.hindborg@wdc.com>