fs: Add LinkFile api support to ZenFS.
This api simulates posix hardlink behavior.
It accepts the source-file and destination-file names, if the source file is
available, it creates an entry in the in-memory files_ map and also
updates the source-file's metadata about the newly linked file and
increments the link count for the source-file and persists this information
on disk.
File deletion ensures that as long as there are links associated to a file,
the file is not deleted on disk and only link specific metadata is updated
and persisted on disk.

Signed-off-by: Aravind Ramesh <aravind.ramesh@wdc.com>