zbd: Fix partial I/O handling in Append and DirectRead
In Zone::Append, the pwrite is changed to use left instead of size as the buffer size.
In ZonedBlockDevice::DirectRead, logic has been changed to the same approach as in other
places in the code, where the number of bytes left is tracked and buffer pointers and
offsets are adjusted with the bytes already read.

Signed-off-by: Jorgen S. Hansen <jorgen.hansen@wdc.com>