zbd: Remove GetXXXFd() uses outside of zbd_zenfs.c
As part of introducing support for ZoneFS for ZBD access, we
need to move away from exposing FDs directly from the
ZonedBlockDevice class. Convert all read operations on
the read and direct read FDs in fs and io to use a new
Read method on ZonedBlockDevice. This call also replaces
existing calls to the DirectRead method.

Signed-off-by: Jorgen Hansen <jorgen.hansen@wdc.com>