zbd: Refactor list zones into a zbdlib specific backend.
Introduce a new ZonedBlockDeviceBackend class that encapsulates
the specific of how a ZBD is accessed, e.g., using zbdlib or
zonefs. The goal is to remove all references to zbdlib structures
from ZonedBlockDevice. This first step introduces the backend
class, and exposes the zones of the device as an opaque object from
the backend.

Signed-off-by: Jorgen Hansen <jorgen.hansen@wdc.com>