zonefs: Get active and open zone limits from sysfs
Starting with Linux 5.19, the max limits for active and open
files are available through sysfs using the paths:
/sys/fs/zonefs/<zbd device name>/max_active_seq_files
/sys/fs/zonefs/<zbd device name>/max_wro_seq_files

To get the device name, /proc/mounts is traversed to identify
the backing device for the zonefs mountpoint.

If we can't determine the backing device, or if the
sysfs entries aren't present, we default to the limits
of a BBZ drive.

Signed-off-by: Jorgen Hansen <jorgen.hansen@wdc.com>