fs: allow files to be deleted while they are being open for write
As we mark up the files as being open for write during migration,
we need to allow deletes of these. There is no reason
to disallow this, as the migrated file will be thrown away
if it has been deleted in the process. It is also consistent
with normal posix file system behavior.

It is safe to keep files open for writing while they have
been deleted by another thread - metadata will not be synced
for that file after it has been marked as deleted and the
space freed up once the open file has been closed.

Signed-off-by: Hans Holmberg <hans.holmberg@wdc.com>