zbd: Move the zbd_open calls into zbdlib backend
Move the zbd_open calls into the zbdlib backend, and make Zone and
ZonedBlockDevice classes access the file descriptors through the
backend using GetXXXFD methods. The GetXXXFD methods of the backend
will be removed at the end of the refactoring, when all FD-ness
has been encapsulated in the backend.

Since zbd_info is obtained through the zbd_open call, the relevant
proporties of the ZBD are now supplied to ZonedBlockDevice through
the backend open call instead of providing the zbdlib specific
zbd_info struct directly to ZonedBlockDevice as part of the zbd_open
call.

Signed-off-by: Jorgen Hansen <jorgen.hansen@wdc.com>