fs: Remove unused GetUniqueID() method from ZoneFile classes
The implementation of GetUniqueID() relies on access to the read FD
for zbd backend. As part of introducing support for ZoneFS for
ZBD access, we need to move away from exposing FDs directly from
the ZonedBlockDevice class. Since GetUniqueID() has no callers,
simply remove it.

FDs cannot be exposed directly anymore because zonefs doesn't present
the zoned block device as a single LBA space. Instead, each zone is
represented as a file, so read/write operations on different zones
translates into read/write operations in different files.

The plan is to continue to present the zoned block device as a single
LBA space, and then (for zonefs) translate that into operations on the
relevant files inside the zonefs specific backend.

Signed-off-by: Jorgen Hansen <jorgen.hansen@wdc.com>